requirements.txt
chardet==3.0.4
click==7.1.2
feedgen==0.9.0
feedparser==6.0.2
Flask==1.1.2
Flask-Login==0.5.0
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
llvmlite==0.34.0
lxml==4.6.1
MarkupSafe==1.1.1
numba==0.51.2
numpy==1.19.3
python-dateutil==2.8.1
sgmllib3k==1.0.0
six==1.15.0
Werkzeug==1.0.1
